+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - added skillsToActivate to onAgentStarted extension event to allow activating skills
 - added noProxy support to network settings
 - replaced floating placement with task, project, and app-level floating placements in UI extensions
+- allowing colons in extension command names
 
 ## [0.70.0]